+++ eggs/IceCreamSandwich/src/main/java/com/android_i/egg/Nyandroid.java
@@ -228,14 +228,14 @@ public class Nyandroid extends Activity {
         mBoard = new Board(this, null);
         setContentView(mBoard);
 
-        mBoard.setOnSystemUiVisibilityChangeListener(new View.OnSystemUiVisibilityChangeListener() {
-            @Override
-            public void onSystemUiVisibilityChange(int vis) {
-                if (0 == (vis & View.SYSTEM_UI_FLAG_HIDE_NAVIGATION)) {
-                    Nyandroid.this.finish();
-                }
-            }
-        });
+//        mBoard.setOnSystemUiVisibilityChangeListener(new View.OnSystemUiVisibilityChangeListener() {
+//            @Override
+//            public void onSystemUiVisibilityChange(int vis) {
+//                if (0 == (vis & View.SYSTEM_UI_FLAG_HIDE_NAVIGATION)) {
+//                    Nyandroid.this.finish();
+//                }
+//            }
+//        });
     }
 
     @Override